# backend/docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Orpind E-Commerce API
  description: |
    Complete REST API for the Orpind premium organic spices e-commerce platform.
    Built with Express.js, MongoDB, and JWT authentication.
  version: 1.0.0
  contact:
    name: Orpind Support
    email: support@orpind.com
  license:
    name: ISC
servers:
  - url: http://localhost:3000/api/v1
    description: Development
  - url: https://api.orpind.com/api/v1
    description: Production
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /orders:
    get:
      tags: [Orders]
      summary: List user orders
      responses:
        '200':
          description: Order list
    post:
      tags: [Orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shippingAddress:
                  type: object
                billingAddress:
                  type: object
                couponCode:
                  type: string
      responses:
        '201':
          description: Order created
...